1. Exit status
added creation of exit status to allow executions to emit failure code, 
this affects command line executions but also e.g. runs via Jenkins jobs 
and allows failed retrieval, deploys or other problems to be raised as a
failing execution or build job run

2. Cache as source repository

the temporary cache folder is now only deleted if artifact coordinates
are supplied, without coordinates it acts as a source repository for 
deployment

+++ maven-repository-provisioner/src/main/java/com/simpligility/maven/provisioner/MavenRepositoryProvisioner.java
@@ -50,12 +50,12 @@ public class MavenRepositoryProvisioner
         }
         catch ( Exception error )
         {
-            logger.info( usage.toString() );
+          logger.info( usage.toString() );
+          exitFailure( "Problem parsing configuration." );
         }
 
         if ( validConfig )
         {
-
             if ( config.getHelp() )
             {
                 logger.info( usage.toString() );
@@ -63,9 +63,14 @@ public class MavenRepositoryProvisioner
                 logger.info( "you can configure the standard Java proxy parameters http.proxyHost, " );
                 logger.info( "http.proxyPort, http.proxyUser and http.proxyPassword. More at " );
                 logger.info( "https://docs.oracle.com/javase/8/docs/api/java/net/doc-files/net-properties.html" );
+                exitFailure( "Invalid configuration." );
             }
             else
             {
+                // overall success of operation, previously was not used so always successful so now we 
+                // use this as default and set to false below as applicable
+                boolean provisioningSuccess = true;
+                StringBuilder provisioningSuccessMessage = new StringBuilder();
                 logger.info( "Provisioning: " + config.getArtifactCoordinate() );
                 logger.info( "Source: " + config.getSourceUrl() );
                 logger.info( "Target: " + config.getTargetUrl() );
@@ -76,13 +81,20 @@ public class MavenRepositoryProvisioner
                 }
                 logger.info( "IncludeSources:" + config.getIncludeSources() );
                 logger.info( "IncludeJavadoc:" + config.getIncludeJavadoc() );
-                logger.info( "Local cache directory: " + config.getCacheDirectory() );
-
+               
+                logger.info( "Local cache or source repository directory: " + config.getCacheDirectory() );
                 cacheDirectory = new File( config.getCacheDirectory() );
                 if ( cacheDirectory.exists() && cacheDirectory.isDirectory() ) 
                 {
-                    logger.info( "Detected local cache directory '" + config.getCacheDirectory() 
-                                 + "' from prior execution." );
+                  logger.info( "Detected local cache directory '" + config.getCacheDirectory() + "'." );
+                  if ( config.getArtifactCoordinate().isEmpty() )
+                  {
+                    logger.info( "No artifact coordinates specified - using cache directory as source." );
+                  } 
+                  else
+                  {
+                    logger.info( "Artifact coordinates specified "
+                        + "- removing stale cache directory from prior execution." );
                     try
                     {
                         FileUtils.deleteDirectory( cacheDirectory );
@@ -91,31 +103,86 @@ public class MavenRepositoryProvisioner
                     catch ( IOException e )
                     {
                         logger.info( config.getCacheDirectory() + " deletion failed" );
+                        exitFailure( "Failed to delete stale cache directory." );
                     }
                     cacheDirectory = new File( config.getCacheDirectory() );
+                  }
+                }
+                ArtifactRetriever retriever = null;
+                if ( !config.getArtifactCoordinate().isEmpty() ) 
+                {
+                  logger.info( "Artifact retrieval starting." );
+                  retriever = new ArtifactRetriever( cacheDirectory );
+                  retriever.retrieve( config.getArtifactCoordinates(), config.getSourceUrl(), 
+                      config.getIncludeSources(), config.getIncludeJavadoc() );
+
+                  logger.info( "Artifact retrieval completed." );
+                } 
+                else 
+                {
+                  logger.info( "Artifact retrieval skipped. " );
                 }
 
-                ArtifactRetriever retriever = new ArtifactRetriever( cacheDirectory );
-                retriever.retrieve( config.getArtifactCoordinates(), config.getSourceUrl(), config.getIncludeSources(),
-                                    config.getIncludeJavadoc() );
-
-                logger.info( "Artifact retrieval completed." );
-
+                logger.info( "Artifact deployment starting." );
                 MavenRepositoryHelper helper = new MavenRepositoryHelper( cacheDirectory );
                 helper.deployToRemote( config.getTargetUrl(), config.getUsername(), config.getPassword(), 
                                        config.getCheckTarget() );
                 logger.info( "Artifact deployment completed." );
 
                 logger.info( "Processing Completed." );
+                StringBuilder summary = new StringBuilder();
+                summary.append( "\nProcessing Summary\n" ).append( DASH_LINE ).append( "\n" );
+                if ( retriever != null )
+                {
+                  summary.append( retriever.listSucessfulRetrievals() ).append( "\n" )
+                    .append( retriever.listFailedTransfers() ).append( "\n" );
 
-                logger.info( "\nProcessing Summary\n"
-                             + DASH_LINE + "\n"
-                             + retriever.listSucessfulRetrievals() + "\n"
-                             + retriever.listFailedTransfers() + "\n"
-                             + helper.listSucessfulDeployments() + "\n"
-                             + helper.listFailedDeployments() + "\n"
-                             + helper.listSkippedDeployment() + "\n" );
+                  if ( retriever.hasFailures() ) 
+                  {
+                    provisioningSuccess = false;
+                    provisioningSuccessMessage.append( retriever.getFailureMessage() );
+                  }
+                  else 
+                  {
+                    provisioningSuccessMessage.append( "Retrieval completed successfully." );
+                  }
+                }
+                
+                summary.append( helper.listSucessfulDeployments() ).append( "\n" )
+                  .append( helper.listFailedDeployments() ).append( "\n" )
+                  .append( helper.listSkippedDeployment() ).append( "\n" );
+                if ( helper.hasFailure() )
+                {
+                  provisioningSuccess = false;
+                  provisioningSuccessMessage.append( helper.getFailureMessage() ); 
+                }
+                else 
+                {
+                  provisioningSuccessMessage.append( "Deployment completed successfully." );
+                }
+                
+                logger.info( summary.toString() );
+                if ( provisioningSuccess ) 
+                {
+                  exitSuccess( provisioningSuccessMessage.toString() );
+                }
+                else
+                {
+                  exitFailure( provisioningSuccessMessage.toString() );
+                }
             }
         }
+    }
+
+    private static void exitSuccess( String message ) 
+    {
+      logger.info( "Exiting: " + message );
+      System.exit( 0 );   
+    }
+
+    private static void exitFailure( String message ) 
+    {
+      logger.info( "Exiting: " + message );
+      System.exit( 1 );
     }
 }
